denied ability for other users to edit and delete other user's items

+++ UI/Controllers/AlbumController.cs
@@ -182,6 +182,13 @@ namespace UI.Controllers
 		[HttpGet]
 		public virtual ActionResult Delete(int id)
 		{
+			var model = _service.GetByID(id, User.Identity.GetUserId());
+			if (model.UserID != User.Identity.GetUserId())
+			{
+				ShowStatusMessage(MessageTypeEnum.error, "This album cannot be deleted by another user", "Delete Failure");
+				return RedirectToAction(MVC.Album.Index());
+			}
+
 			_service.Delete(id, User.Identity.GetUserId());
 
 			ShowStatusMessage(MessageTypeEnum.success, "", "Album Deleted Successfully");
