+++ UI/Controllers/GameController.cs
@@ -85,6 +85,11 @@ namespace UI.Controllers
 		{
 			ViewBag.Title = "Edit";
 			var model = _service.GetByID(id, User.Identity.GetUserId());
+			if (model.UserID != User.Identity.GetUserId())
+			{
+				ShowStatusMessage(MessageTypeEnum.error, "This game cannot be edited by another user", "Edit Failure");
+				return RedirectToAction(MVC.Game.Index());
+			}
 
 			return View(model);
 		}
@@ -117,6 +122,13 @@ namespace UI.Controllers
 		[HttpGet]
 		public virtual ActionResult Delete(int id)
 		{
+			var model = _service.GetByID(id, User.Identity.GetUserId());
+			if (model.UserID != User.Identity.GetUserId())
+			{
+				ShowStatusMessage(MessageTypeEnum.error, "This game cannot be deleted by another user", "Delete Failure");
+				return RedirectToAction(MVC.Game.Index());
+			}
+
 			_service.Delete(id, User.Identity.GetUserId());
 
 			ShowStatusMessage(MessageTypeEnum.success, "", "Game Deleted Successfully");
