+++ UI/Controllers/MovieController.cs
@@ -121,6 +121,13 @@ namespace UI.Controllers
 		[HttpGet]
 		public virtual ActionResult Delete(int id)
 		{
+			var model = _service.GetByID(id, User.Identity.GetUserId());
+			if (model.UserID != User.Identity.GetUserId())
+			{
+				ShowStatusMessage(MessageTypeEnum.error, "This game cannot be deleted by another user", "Delete Failure");
+				return RedirectToAction(MVC.Game.Index());
+			}
+
 			_service.Delete(id, User.Identity.GetUserId());
 
 			ShowStatusMessage(MessageTypeEnum.success, "", "Movie Deleted Successfully");
